pyproject.toml
[project]
name = "HWdjango2"
version = "0.1.0"
description = ""
authors = [
    {name = "LRK",email = "kirillvalentinka@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.0.0",
    "djangorestframework>=3.14.0",
    "Pillow>=10.0.0",
    "django-filter>=25.0.0",
    "djangorestframework-simplejwt>=5.0.0",
    "drf-spectacular>=0.29.0",
    "stripe>=14.0.0",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
